Moved services as endpoint implementations.
