--- Application/Common/Interfaces/ICurrentUserService.cs
@@ -1,6 +1,0 @@
-namespace Wallet.Application.Common.Interfaces;
-
-public interface ICurrentUserService
-{
-    string UserId { get; }
-}
